% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hampel.R
\name{Hampel}
\alias{Hampel}
\title{Robust Mean by Hampel Estimation}
\usage{
Hampel(df, a = 1.5, b = 3, c = 4.5, s_star = Q_method(df))
}
\arguments{
\item{df}{data frame containing one row per lab, and one column per replicate.}

\item{a, b, c}{tuning parameters a<b<c. Greater efficiency is obtained by
increasing the range; improved resistance to outliers or minor modes is
obtained by reducing the range.}

\item{s_star}{robust standard deviation used in calculations. Defaults to the
Q method.}
}
\value{
robust mean
}
\description{
This function calculates robust mean with the 'Hampel estimator' described in
ISO 13528-2015, section C.5.3.3
}
\examples{
Hampel(data.frame(replicate1=c(1,2,3,4,5),replicate2 = c(6,7,NA,100,NA)))
}
