% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homogeneity.R
\name{Homogeneity}
\alias{Homogeneity}
\title{Homogeneity by between-sample standard deviation estimate}
\usage{
Homogeneity(df)
}
\arguments{
\item{df}{gx2 data.frame, where g is the number of labs participating in the
test.}
}
\value{
estimate of between-sample standard deviation
}
\description{
This function calculates measures homogeneity as described in
ISO 13528-2015, section B.3, page 48.
}
\examples{
Homogeneity(data.frame(replicate1=c(1,2,3,4,5),replicate2 = c(6,7,3,4,6)))
}
