% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_kernel_density.R
\name{Plot_kernel_density}
\alias{Plot_kernel_density}
\title{Kernel Density Plot}
\usage{
Plot_kernel_density(x, kernal_sd = "a")
}
\arguments{
\item{x}{a numerical vector of data}

\item{kernal_sd}{a character giving the method for determining kernel density
standard deviation (default: 'a' according to ISO 13528-2015, section 10.3.i.a,
where s* is calculated by Algorithm_A(x)) or numerical standard deviation.}
}
\value{
ggplot object
}
\description{
This function returns a kernel density plot (ggplot) as described in
ISO 13528-2015, section 10.3
}
\examples{
Plot_kernel_density(c(0.2640, 0.2670, 0.2960 ,0.3110, 0.3310, 0.4246))
Plot_kernel_density(c(0.2640, 0.2670, 0.2960 ,0.3110, 0.3310, 0.4246),0.025)
}
