% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_Hampel.R
\name{Q_Hampel}
\alias{Q_Hampel}
\title{Robust Mean and Standard Deviation by Q/Hampel-method}
\usage{
Q_Hampel(df, a = 1.5, b = 3, c = 4)
}
\arguments{
\item{df}{data frame containing one row per lab, and one column per replicate.}

\item{a, b, c}{tuning parameters a<b<c. Greater efficiency is obtained by
increasing the range; improved resistance to outliers or minor modes is
obtained by reducing the range.}
}
\value{
robust mean
}
\description{
This function calculates robust mean and standard deviation with the
'Q/Hampel method' described in ISO 13528-2015, section C.5.4
}
\examples{
Q_Hampel(data.frame(replicate1=c(1,2,3,4,5),replicate2 = c(6,7,NA,100,NA)))
}
