% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_method.R
\name{Q_method}
\alias{Q_method}
\title{Robust Mean and Standard Deviation}
\usage{
Q_method(df)
}
\arguments{
\item{df}{data frame containing one row per lab, and one column per replicate.}
}
\value{
robust standard deviation
}
\description{
This function calculates robust SD with the 'Q method' described in
ISO 13528-2015, section C.5.2.2
}
\examples{
Q_method(data.frame(replicate1=c(1,2,3,4,5),replicate2 = c(6,7,NA,100,NA)))
}
